requirements.txt
beautifulsoup4
lxml
google-cloud-vision
requests
pillow
python-dotenv
svgpathtools
google-generativeai